requirements.txt

flask
openai
faiss-cpu
numpy
PyPDF2
python-docx
spacy
